requirements.txt
streamlit==1.37.1
pandas==2.2.2
numpy==1.26.4
python-docx==1.1.2
openpyxl==3.1.5
